pyproject.toml
[project]
name = "task-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt==3.2.2",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.24",
    "sqlalchemy[asyncio]>=2.0.48",
    "structlog>=25.5.0",
    "testcontainers>=4.14.2",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "alembic>=1.18.4",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
